pyproject.toml
[tool.poetry]
name = "oxebot"
version = "0.1.0"
description = ""
authors = ["yzakius <dev@yzakius.me>"]

[tool.poetry.dependencies]
python = "^3.8"
python-decouple = "^3.3"
requests = "^2.24.0"
pyTelegramBotAPI = "^3.7.2"
flake8 = "^3.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
